fix: Finalize SkidKZ prototype with Russian UI, stable navigation, and mock economy

+++ skidkz/lib/features/buyer/screens/product_detail_screen.dart
@@ -180,7 +180,7 @@ class _ProductDetailScreenState extends ConsumerState<ProductDetailScreen> {
         buyerId: ref.read(authProvider)!.id,
         sellerId: product.sellerId,
         product: product,
-        amount: product.skidkzPrice,
+        amount: product.skidkzPrice.toDouble(),
         promoCode: _appliedPromo,
         status: OrderStatus.paid,
         createdAt: DateTime.now(),
